requirements.txt
altair==5.5.0
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.1
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
docker==7.1.0
entrypoints==0.4
et_xmlfile==2.0.0
extra-streamlit-components==0.1.71
extra-streamlit-components-better-cookie-manager==0.0.10
Faker==37.1.0
favicon==0.7.0
fonttools==4.56.0
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
htbuilder==0.9.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mysql-connector-python==9.2.0
narwhals==1.26.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
prometheus_client==0.21.1
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.3
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
soupsieve==2.6
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.47.0
streamlit-authenticator==0.4.2
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-cookies-controller==0.0.4
streamlit-cookies-manager==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-pdf-viewer==0.0.22
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_pdf_reader==0.0.5
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
xlwings==0.33.11
